fix: Resolve circular import
